add test suite and user seeds for travis ci
